% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{write_attributable_deaths}
\alias{write_attributable_deaths}
\title{Write outputs to csv}
\usage{
write_attributable_deaths(
  df_list,
  matsim,
  arraysim,
  totdeath,
  attrdl_yr_all,
  attr_fractions_yr,
  output_folder_path = NULL
)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{matsim}{A matrix (numeric). Total (glob),
cold and heat-attributable deaths per region from reduced coefficients.}

\item{arraysim}{An array (numeric). Total (glob),
cold and heat-attributable deaths per region for 1000 simulations.
Used to derive confidence intervals.}

\item{totdeath}{A named vector of integers. Total observed mortality per region.}

\item{attrdl_yr_all}{A dataframe with attributable deaths by year for each region.}

\item{attr_fractions_yr}{A dataframe with attributable fractions by year and region.}

\item{output_folder_path}{Path to folder for storing outputs.}
}
\value{
\itemize{
\item \code{anregions_bind} A matrix of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals), disaggregated by region.
\item \code{antot_bind} A matrix of numbers of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals).
\item \code{afregions_bind} A matrix of fractions of all-cause mortality
attributable to temperature, heat, cold, extreme heat and extreme cold
(with confidence intervals), disaggregated by region.
\item \code{aftot_bind} A matrix of fractions of all-cause mortality
attributable to temperature, heat, cold, extreme heat and extreme cold
(with confidence intervals).
}
}
\description{
Write the attributable deaths and temperature for each regions,
with empirical CI estimated using the re-centered bases.
}
\examples{
output_folder_path = 'myfolder/output/'
}
